Teachers Page Half Done

+++ src/main/java/com/example/schoolmanagement/Admin.java
@@ -4,13 +4,13 @@ public class Admin extends Employee{
     private final String username;
     private final String password;
 
-    public Admin(String name, String email, String gender, double salary, String username, String password) {
+    public Admin(String name, String email, String gender, float salary, String username, String password) {
         super(name, email, gender, salary);
         this.username = username;
         this.password = password;
     }
 
-    public Admin(int id, String name, String email, String gender, double salary, String username, String password) {
+    public Admin(int id, String name, String email, String gender, float salary, String username, String password) {
         super(id, name, email, gender, salary);
         this.username = username;
         this.password = password;
