+++ src/main/java/com/example/schoolmanagement/Employee.java
@@ -3,16 +3,16 @@ package com.example.schoolmanagement;
 import java.util.Date;
 
 public class Employee extends Person {
-    private double salary;
+    private float salary;
     private final Date hireDate;
 
-    public Employee(String name, String email, String gender, double salary) {
+    public Employee(String name, String email, String gender, float salary) {
         super(name, email, gender);
         this.salary = salary;
         this.hireDate = new Date();
     }
 
-    public Employee(int id, String name, String email, String gender, double salary) {
+    public Employee(int id, String name, String email, String gender, float salary) {
         super(id, name, email, gender);
         this.salary = salary;
         this.hireDate = new Date();
@@ -22,7 +22,7 @@ public class Employee extends Person {
         return salary;
     }
 
-    public void setSalary(double salary) {
+    public void setSalary(float salary) {
         this.salary = salary;
     }
